storage/generic: move resuable tests into separate files
In order to be able to import these tests in the future, we need them to
not be in `_test.go` files.
